fix(carrinho): :sparkles: Adicionado componente produto_carrinho e corrigido caminho dos arquivos
